Raspi/requirements.txt
# Surveillance Car - Complete Requirements
# Hardware Control, MQTT Communication, and AI Vision

# ============================================
# Hardware Control
# ============================================
RPi.GPIO>=0.7.1
PyYAML>=6.0.1

# ============================================
# Computer Vision & Camera
# ============================================
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# ============================================
# MQTT Communication
# ============================================
paho-mqtt>=1.6.1

# ============================================
# WebSocket Streaming
# ============================================
websockets>=12.0
pyaudio>=0.2.13

# ============================================
# Testing
# ============================================
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# ============================================
# Development Tools (Optional)
# ============================================
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
